requirements.txt
mysql_connector_repackaged==0.3.1
pymongo==4.2.0
